\name{mogene10stv174probe}
\docType{data}
\alias{mogene10stv174probe}
\title{Probe sequence for Affymetrix MoGene microarrays.}
\description{This data object contains probe sequences for all probes/probesets defined in the mogene10stv174cdf CDF file/package version 0.3.0.}
\usage{data(mogene10stv174probe)}
\format{A data frame with xx rows and 6 columns, as follows.
  \tabular{lll}{
    sequence \tab character \tab probe sequence \cr
    x  \tab integer  \tab x-coordinate on the array \cr
    y  \tab integer  \tab y-coordinate on the array \cr
    Probe.Set.Name      \tab character \tab Affymetrix Probe Set Name \cr
    Probe.Interrogation.Position \tab integer \tab Probe Interrogation Position \cr
    Target.Strandedness \tab factor \tab Target Strandedness\cr
}}

\source{The probe sequence data was provided by Affymetrix. The probe file was generated by the make_cdf_from_db.pl perl function for the generation of custom/remapped CDF files for Affymetrix Gene ST arrays.}

\keyword{datasets}
\examples{
  data( mogene10stv174probe )
  as.data.frame(mogene10stv174probe[1:3,])
}
